details button

+++ daas/daas_app/models.py
@@ -2,7 +2,7 @@ from django.db import models
 
 
 class Statistics(models.Model):
-    timeout = models.FloatField(default=None, blank=True, null=True)
+    timeout = models.IntegerField(default=None, blank=True, null=True)
     elapsed_time = models.IntegerField(default=None, blank=True, null=True)
     exit_status = models.IntegerField(default=None, blank=True, null=True)
     timed_out = models.BooleanField(default=False)
